Remove explicit config.boot reference in confmode.cc process()
Part of VRVDR-49437

config.boot will no longer exist as the boot config file once VRVDR-32174 is fully implemented (removal of config.boot).  If a config.boot file is present, it will be used, but it is no longer the default file, and 'save config.boot' is not the equivalent of 'save' with no file specified (save is implicit on 'commit' operations, so 'save' is in fact a no-op, and the file used by 'commit' will not be 'config.boot' any longer).

So, I am removing the special case for 'action == "save config.boot" so it is no longer treated the same as 'save' and instead is now an explicit request to save the config as 'config.boot'.