Transform lighttpd cron with dh_configpackage
Use dh_configpackage and a transform operation to apply DANOS tweaks
to the /etc/cron.daily/lighttpd script. These tweaks were previously
performed by another package not present in DANOS.

The tweaks prevent the cron script from attempting to clean the
lighttpd cache directory if they are empty, as would be the case if
the REST API has not been configured.

The run-parts-cron-filename-contains-illegal-chars Lintian warning is
generated by this change since we are installing the transformed
script as /etc/cron.daily/lighttpd.vyatta and as such will not be
executed by run-parts during a cron run. This is intentional since on
package install /etc/cron.daily/lighttpd.vyatta will be symlinked to
/etc/cron.daily/lighttpd - which will be executed by run-parts.

Since the generation of lighttpd.vyatta happens at build time a build
dependency on lighttpd is added, so that /etc/cron.daily/lighttpd is
in the build environment.